make cam-fv/model_mod work correctly for VERTISUNDEF observations.
fix a remaining bug in cam model_mod related to assimilating obs
with a vertical type of VERTISUNDEF.  the vertical conversion
code went through a path that didn't set the return status code,
which usually resulted in the obs being discarded from the close list.

also make the quad interp code use the 'missing values are ok in the state'
function to see if the interpolation should test for MISSING_R8 at
any of the quad corners before doing interpolation.  the code was
already there but was hardcoded either in or out.  now it is based on
the setting in the options module at runtime.