Update index.html
Changed all references of $DART to $DART_DIR and of $WORKDIR to $BASE_DIR so that the environmental variables declared in the tutorial match the variables in param.csh. The hope is that this makes the tutorial easier to understand and so that followers of the tutorial do not make $BASE_DIR the same directory as their $DART_DIR/models/wrf/work directory. Also, removed the line about editing input.nml to alter ens_size since ens_size is already set by the sed invocation in setup.csh.