reference docs directory. removed svn baggage
The documentation director was renamed to docs a long time ago.
The svn junk at the bottom of the files is moot (and wrong), so I removed it.
The link to the documention main index is still wrong, but at least consistent
so that when we know what to reference, changing it should be easy.