fix confusing error with 'no update' variables and inflation files
changed fill_inflation_restart so it overwrites the status of the
'update' flag to always be true before writing out the inflation files.
this ensures that even 'no update' variables will have the required
inflation variables in the restart files.

to use these files, copy the inputs to the outputs in the script
before running filter, and filter will update the inflation values
for only the 'update' fields, but the original variables for 'no update'
will still exist (and be unchanged) for the next cycle.

this fixes a long-standing problem with 'no update' variables.
it's tim's idea and it's a good one.