removing execute permission and mkmf.template
The mkmf.template was ALWAYS changed and is just a colossal nuisance.
Since the user is supposed to pick an appropriate one and copy it
into place - they should have one.  If it is part of the .gitignore
it will not normally show up in a 'git status' and will stay put
when changing branches, etc.